fix: break into multiple stories and change structure
