+++ src/DescriptionList/stories/DescriptionList.story.tsx
@@ -1,0 +1,9 @@
+import DescriptionList from "../DescriptionList";
+
+export default {
+  title: "Components/DescriptionList",
+  component: DescriptionList,
+};
+
+export { GroupMinWidth } from "./DescriptionList.group-min-width.story";
+export { Playground } from "./DescriptionList.playground.story";
