+++ src/TopBar/index.ts
@@ -1,1 +1,2 @@
 export { TopBar } from "./TopBar";
+export type { MenuItemLinkProps } from "./components/MenuItemLink";
